split commponents
