+++ client/src/components/Chat/Header.js
@@ -1,7 +1,6 @@
 import React, { useState } from "react";
 import {
   Avatar,
-  Badge,
   Box,
   Divider,
   Fade,
@@ -9,7 +8,6 @@ import {
   Menu,
   MenuItem,
   Stack,
-  styled,
   Typography,
 } from "@mui/material";
 import { useSearchParams } from "react-router-dom";
@@ -17,35 +15,7 @@ import { CaretDown, MagnifyingGlass, Phone, VideoCamera } from "phosphor-react";
 import { faker } from "@faker-js/faker";
 import { useTheme } from "@mui/material/styles";
 import useResponsive from "./../../hooks/useResponsive";
-
-const StyledBadge = styled(Badge)(({ theme }) => ({
-  "& .MuiBadge-badge": {
-    backgroundColor: "#44b700",
-    color: "#44b700",
-    boxShadow: `0 0 0 2px ${theme.palette.background.paper}`,
-    "&::after": {
-      position: "absolute",
-      top: 0,
-      left: 0,
-      width: "100%",
-      height: "100%",
-      borderRadius: "50%",
-      animation: "ripple 1.2s infinite ease-in-out",
-      border: "1px solid currentColor",
-      content: '""',
-    },
-  },
-  "@keyframes ripple": {
-    "0%": {
-      transform: "scale(.8)",
-      opacity: 1,
-    },
-    "100%": {
-      transform: "scale(2.4)",
-      opacity: 0,
-    },
-  },
-}));
+import StyledBadge from "../StyledBadge";
 
 const Conversation_Menu = [
   {
@@ -70,9 +40,11 @@ const Header = () => {
   const [conversationMenuAnchorEl, setConversationMenuAnchorEl] =
     useState(null);
   const openConversationMenu = Boolean(conversationMenuAnchorEl);
+
   const handleClickConversationMenu = (event) => {
     setConversationMenuAnchorEl(event.currentTarget);
   };
+
   const handleCloseConversationMenu = () => {
     setConversationMenuAnchorEl(null);
   };
@@ -164,8 +136,8 @@ const Header = () => {
           >
             <Box p={1}>
               <Stack spacing={1}>
-                {Conversation_Menu.map((el) => (
-                  <MenuItem onClick={handleCloseConversationMenu}>
+                {Conversation_Menu.map((el, idx) => (
+                  <MenuItem onClick={handleCloseConversationMenu} key={idx}>
                     <Stack
                       sx={{ minWidth: 100 }}
                       direction="row"
